requirements.txt
streamlit
pytesseract
pillow
numpy
pandas
scikit-learn
